+++ Editor/Unity.Build.Classic.Private/BuildPipelineSelector.cs
@@ -1,14 +1,28 @@
-﻿using System.Linq;
+﻿#if ENABLE_EXPERIMENTAL_INCREMENTAL_PIPELINE
+using Unity.Build.Classic.Private.IncrementalClassicPipeline;
+#endif
 using Unity.BuildSystem.NativeProgramSupport;
 
 namespace Unity.Build.Classic.Private
 {
     class BuildPipelineSelector : BuildPipelineSelectorBase
     {
-        public override BuildPipelineBase SelectFor(Platform platform)
+        public override BuildPipelineBase SelectFor(Platform platform, bool incremental)
         {
-            var classicPipelineBases = TypeCacheHelper.ConstructTypesDerivedFrom<ClassicPipelineBase>();
-            return classicPipelineBases.FirstOrDefault(pipeline => pipeline.Platform.GetType() == platform.GetType());
+            if (platform == null)
+            {
+                return null;
+            }
+#if ENABLE_EXPERIMENTAL_INCREMENTAL_PIPELINE
+            if (incremental)
+            {
+                return TypeCacheHelper.ConstructTypeDerivedFrom<ClassicIncrementalPipelineBase>(p => p.Platform.GetType() == platform.GetType());
+            }
+            else
+#endif
+            {
+                return TypeCacheHelper.ConstructTypeDerivedFrom<ClassicNonIncrementalPipelineBase>(p => p.Platform.GetType() == platform.GetType());
+            }
         }
     }
 }