+++ Editor/Unity.Build/Common/PlayerScriptingDefines.cs
@@ -1,8 +1,10 @@
 using System;
 using Unity.Properties;
+using Unity.Serialization;
 
 namespace Unity.Build.Common
 {
+    [FormerName("Unity.Build.Common.PlayerScriptingDefines, Unity.Build.Common")]
     public sealed class PlayerScriptingDefines : IBuildComponent
     {
         [CreateProperty]