+++ Editor/Unity.Build/Common/ScriptingDebuggerSettings.cs
@@ -1,7 +1,9 @@
 using Unity.Properties;
+using Unity.Serialization;
 
 namespace Unity.Build.Common
 {
+    [FormerName("Unity.Build.Common.ScriptingDebuggerSettings, Unity.Build.Common")]
     public sealed class ScriptingDebuggerSettings : IBuildComponent
     {
         [CreateProperty]